+++ models/node_reform.go
@@ -47,7 +47,7 @@ func (v *nodeTableType) PKColumnIndex() uint {
 
 // NodeTable represents nodes view or table in SQL database.
 var NodeTable = &nodeTableType{
-	s: parse.StructInfo{Type: "Node", SQLSchema: "", SQLName: "nodes", Fields: []parse.FieldInfo{{Name: "ID", PKType: "int64", Column: "id"}, {Name: "Type", PKType: "", Column: "type"}, {Name: "Name", PKType: "", Column: "name"}}, PKFieldIndex: 0},
+	s: parse.StructInfo{Type: "Node", SQLSchema: "", SQLName: "nodes", Fields: []parse.FieldInfo{{Name: "ID", PKType: "int32", Column: "id"}, {Name: "Type", PKType: "", Column: "type"}, {Name: "Name", PKType: "", Column: "name"}}, PKFieldIndex: 0},
 	z: new(Node).Values(),
 }
 
@@ -110,9 +110,9 @@ func (s *Node) HasPK() bool {
 // SetPK sets record primary key.
 func (s *Node) SetPK(pk interface{}) {
 	if i64, ok := pk.(int64); ok {
-		s.ID = int64(i64)
+		s.ID = int32(i64)
 	} else {
-		s.ID = pk.(int64)
+		s.ID = pk.(int32)
 	}
 }
 
@@ -162,7 +162,7 @@ func (v *rDSNodeTableType) PKColumnIndex() uint {
 
 // RDSNodeTable represents nodes view or table in SQL database.
 var RDSNodeTable = &rDSNodeTableType{
-	s: parse.StructInfo{Type: "RDSNode", SQLSchema: "", SQLName: "nodes", Fields: []parse.FieldInfo{{Name: "ID", PKType: "int64", Column: "id"}, {Name: "Type", PKType: "", Column: "type"}, {Name: "Name", PKType: "", Column: "name"}, {Name: "Region", PKType: "", Column: "region"}}, PKFieldIndex: 0},
+	s: parse.StructInfo{Type: "RDSNode", SQLSchema: "", SQLName: "nodes", Fields: []parse.FieldInfo{{Name: "ID", PKType: "int32", Column: "id"}, {Name: "Type", PKType: "", Column: "type"}, {Name: "Name", PKType: "", Column: "name"}, {Name: "Region", PKType: "", Column: "region"}}, PKFieldIndex: 0},
 	z: new(RDSNode).Values(),
 }
 
@@ -228,9 +228,9 @@ func (s *RDSNode) HasPK() bool {
 // SetPK sets record primary key.
 func (s *RDSNode) SetPK(pk interface{}) {
 	if i64, ok := pk.(int64); ok {
-		s.ID = int64(i64)
+		s.ID = int32(i64)
 	} else {
-		s.ID = pk.(int64)
+		s.ID = pk.(int32)
 	}
 }
